refactor: update pages with eslint rules
