+++ src/pages/main/ui/Page.tsx
@@ -2,7 +2,7 @@ import LatestNews from "./LatestNews/LatestNews";
 import NewsByFilters from "./NewsByFilters/NewsByFilters";
 import styles from "./styles.module.css";
 
-const MainPage = () => {
+export const MainPage = () => {
   return (
     <main className={styles.main}>
       <LatestNews />
@@ -11,5 +11,3 @@ const MainPage = () => {
     </main>
   );
 };
-
-export default MainPage;
