+++ src/pages/news/ui/index.ts
@@ -1,0 +1,1 @@
+export { NewsPage } from "./Page";
